--- src/components/side_bar/SideBar.css
@@ -1,105 +1,0 @@
-.SideBar-Style{
-    width: 250px;
-    background-color: #2f2f2f;
-    transition: width 0.2s ease;
-    flex-direction: column;
-}
-
-.SideBar-Style.open{
-    width: 250px;
-}
-
-.SideBar-Style.closed{
-    width: 50px;
-}
-
-.C{
-    display: flex;
-    flex-direction: row;
-}
-.Hamburger-Style{
-    display: flex;
-    color: white;
-    background-color: transparent;
-}
-
-/* 新規作成ボタン*/
-.New-Note-Button{
-    display: flex;
-    width: 100%;
-    background-color: transparent;
-}
-
-/* 箱 */
-.Contents-and-Delete{
-    display: flex;
-    width: 100%;
-    background-color: transparent;
-    align-items: center;
-}
-
-/* キャンセルボタン */
-.Cancel-Button{
-    background-color: gray;
-    color: white;
-    font-size: small;
-}
-
-/* ノート一覧(文字)*/
-.Contents-Name{
-    text-align: center;
-    flex: 1;
-}
-
-/* 削除ボタン　通常時 */
-.Delete-Button{
-    size: small;
-    background-color: transparent;
-}
-
-/* 削除ボタン　削除モード時 */
-.Delete-Button.on{
-    border:2px solid darkred;
-    color: white;
-}
-
-/* 一つ一つのノートボタン*/
-.Notes-Style{
-    width: 100%;
-    height: 5vh;
-    background-color: transparent;
-    color: white;
-    font-size: small;
-}
-
-/* 選択されているノート*/
-.Notes-Style.selected{
-    border: 1px solid #5ac8fa;
-}
-
-/* 削除モードの時 */
-.Notes-Style.Delete-Mode.selected{
-    border: 1px solid red;
-}
-
-.A-Note-Style{
-    display: flex;
-    justify-content: space-between;
-}
-
-.A-Note-Title-Style{
-    display: flex;
-    align-items: center;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.A-Note-Date-Style{
-    font-size: small;
-}
-
-.SideBar-Style ul {
-    list-style: none;  /* 丸を消す */
-    padding-left: 0;   /* 左の余白も消す */
-    margin: 0;
-}
